% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Context}
\alias{Context}
\title{CustomSearch API Objects 
Lets you search over a website or collection of websites}
\usage{
Context(facets = NULL, title = NULL)
}
\arguments{
\item{facets}{No description}

\item{title}{No description}
}
\value{
Context object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:04:50
filename: /Users/mark/dev/R/autoGoogleAPI/googlecustomsearchv1.auto/R/customsearch_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Context Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

