% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_functions.R
\docType{package}
\name{customsearch_googleAuthR}
\alias{customsearch_googleAuthR}
\alias{customsearch_googleAuthR-package}
\title{CustomSearch API
Lets you search over a website or collection of websites}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:41:21
filename: /Users/mark/dev/R/autoGoogleAPI/googlecustomsearchv1.auto/R/customsearch_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
