% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Promotion.bodyLines}
\alias{Promotion.bodyLines}
\title{Promotion.bodyLines Object}
\usage{
Promotion.bodyLines()
}
\value{
Promotion.bodyLines object
}
\description{
Promotion.bodyLines Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Promotion functions: \code{\link{Promotion.image}},
  \code{\link{Promotion}}
}

