% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Promotion.image}
\alias{Promotion.image}
\title{Promotion.image Object}
\usage{
Promotion.image(height = NULL, source = NULL, width = NULL)
}
\arguments{
\item{height}{No description}

\item{source}{No description}

\item{width}{No description}
}
\value{
Promotion.image object
}
\description{
Promotion.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Promotion functions: \code{\link{Promotion.bodyLines}},
  \code{\link{Promotion}}
}

