% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Promotion}
\alias{Promotion}
\title{Promotion Object}
\usage{
Promotion(Promotion.bodyLines = NULL, Promotion.image = NULL,
  bodyLines = NULL, displayLink = NULL, htmlTitle = NULL, image = NULL,
  link = NULL, title = NULL)
}
\arguments{
\item{Promotion.bodyLines}{The \link{Promotion.bodyLines} object or list of objects}

\item{Promotion.image}{The \link{Promotion.image} object or list of objects}

\item{bodyLines}{No description}

\item{displayLink}{No description}

\item{htmlTitle}{No description}

\item{image}{No description}

\item{link}{No description}

\item{title}{No description}
}
\value{
Promotion object
}
\description{
Promotion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Promotion functions: \code{\link{Promotion.bodyLines}},
  \code{\link{Promotion.image}}
}

