% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Query}
\alias{Query}
\title{Query Object}
\usage{
Query(count = NULL, cr = NULL, cref = NULL, cx = NULL,
  dateRestrict = NULL, disableCnTwTranslation = NULL, exactTerms = NULL,
  excludeTerms = NULL, fileType = NULL, filter = NULL, gl = NULL,
  googleHost = NULL, highRange = NULL, hl = NULL, hq = NULL,
  imgColorType = NULL, imgDominantColor = NULL, imgSize = NULL,
  imgType = NULL, inputEncoding = NULL, language = NULL,
  linkSite = NULL, lowRange = NULL, orTerms = NULL,
  outputEncoding = NULL, relatedSite = NULL, rights = NULL, safe = NULL,
  searchTerms = NULL, searchType = NULL, siteSearch = NULL,
  siteSearchFilter = NULL, sort = NULL, startIndex = NULL,
  startPage = NULL, title = NULL, totalResults = NULL)
}
\arguments{
\item{count}{No description}

\item{cr}{No description}

\item{cref}{No description}

\item{cx}{No description}

\item{dateRestrict}{No description}

\item{disableCnTwTranslation}{No description}

\item{exactTerms}{No description}

\item{excludeTerms}{No description}

\item{fileType}{No description}

\item{filter}{No description}

\item{gl}{No description}

\item{googleHost}{No description}

\item{highRange}{No description}

\item{hl}{No description}

\item{hq}{No description}

\item{imgColorType}{No description}

\item{imgDominantColor}{No description}

\item{imgSize}{No description}

\item{imgType}{No description}

\item{inputEncoding}{No description}

\item{language}{No description}

\item{linkSite}{No description}

\item{lowRange}{No description}

\item{orTerms}{No description}

\item{outputEncoding}{No description}

\item{relatedSite}{No description}

\item{rights}{No description}

\item{safe}{No description}

\item{searchTerms}{No description}

\item{searchType}{No description}

\item{siteSearch}{No description}

\item{siteSearchFilter}{No description}

\item{sort}{No description}

\item{startIndex}{No description}

\item{startPage}{No description}

\item{title}{No description}

\item{totalResults}{No description}
}
\value{
Query object
}
\description{
Query Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
