% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Result.image}
\alias{Result.image}
\title{Result.image Object}
\usage{
Result.image(byteSize = NULL, contextLink = NULL, height = NULL,
  thumbnailHeight = NULL, thumbnailLink = NULL, thumbnailWidth = NULL,
  width = NULL)
}
\arguments{
\item{byteSize}{No description}

\item{contextLink}{No description}

\item{height}{No description}

\item{thumbnailHeight}{No description}

\item{thumbnailLink}{No description}

\item{thumbnailWidth}{No description}

\item{width}{No description}
}
\value{
Result.image object
}
\description{
Result.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Result functions: \code{\link{Result.labels}},
  \code{\link{Result.pagemap}}, \code{\link{Result}}
}

