% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Result.labels}
\alias{Result.labels}
\title{Result.labels Object}
\usage{
Result.labels()
}
\value{
Result.labels object
}
\description{
Result.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Result functions: \code{\link{Result.image}},
  \code{\link{Result.pagemap}}, \code{\link{Result}}
}
