% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Result.pagemap}
\alias{Result.pagemap}
\title{Result.pagemap Object}
\usage{
Result.pagemap()
}
\value{
Result.pagemap object
}
\description{
Result.pagemap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Result functions: \code{\link{Result.image}},
  \code{\link{Result.labels}}, \code{\link{Result}}
}
