% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Result}
\alias{Result}
\title{Result Object}
\usage{
Result(Result.image = NULL, Result.labels = NULL, Result.pagemap = NULL,
  cacheId = NULL, displayLink = NULL, fileFormat = NULL,
  formattedUrl = NULL, htmlFormattedUrl = NULL, htmlSnippet = NULL,
  htmlTitle = NULL, image = NULL, labels = NULL, link = NULL,
  mime = NULL, pagemap = NULL, snippet = NULL, title = NULL)
}
\arguments{
\item{Result.image}{The \link{Result.image} object or list of objects}

\item{Result.labels}{The \link{Result.labels} object or list of objects}

\item{Result.pagemap}{The \link{Result.pagemap} object or list of objects}

\item{cacheId}{No description}

\item{displayLink}{No description}

\item{fileFormat}{No description}

\item{formattedUrl}{No description}

\item{htmlFormattedUrl}{No description}

\item{htmlSnippet}{No description}

\item{htmlTitle}{No description}

\item{image}{No description}

\item{labels}{No description}

\item{link}{No description}

\item{mime}{No description}

\item{pagemap}{No description}

\item{snippet}{No description}

\item{title}{No description}
}
\value{
Result object
}
\description{
Result Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Result functions: \code{\link{Result.image}},
  \code{\link{Result.labels}}, \code{\link{Result.pagemap}}
}

