% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Search.queries}
\alias{Search.queries}
\title{Search.queries Object}
\usage{
Search.queries()
}
\value{
Search.queries object
}
\description{
Search.queries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Search functions: \code{\link{Search.searchInformation}},
  \code{\link{Search.spelling}}, \code{\link{Search.url}},
  \code{\link{Search}}
}

