% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Search}
\alias{Search}
\title{Search Object}
\usage{
Search(Search.queries = NULL, Search.searchInformation = NULL,
  Search.spelling = NULL, Search.url = NULL, context = NULL,
  items = NULL, promotions = NULL, queries = NULL,
  searchInformation = NULL, spelling = NULL, url = NULL)
}
\arguments{
\item{Search.queries}{The \link{Search.queries} object or list of objects}

\item{Search.searchInformation}{The \link{Search.searchInformation} object or list of objects}

\item{Search.spelling}{The \link{Search.spelling} object or list of objects}

\item{Search.url}{The \link{Search.url} object or list of objects}

\item{context}{No description}

\item{items}{No description}

\item{promotions}{No description}

\item{queries}{No description}

\item{searchInformation}{No description}

\item{spelling}{No description}

\item{url}{No description}
}
\value{
Search object
}
\description{
Search Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Search functions: \code{\link{Search.queries}},
  \code{\link{Search.searchInformation}},
  \code{\link{Search.spelling}}, \code{\link{Search.url}}
}
