% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Search.searchInformation}
\alias{Search.searchInformation}
\title{Search.searchInformation Object}
\usage{
Search.searchInformation(formattedSearchTime = NULL,
  formattedTotalResults = NULL, searchTime = NULL, totalResults = NULL)
}
\arguments{
\item{formattedSearchTime}{No description}

\item{formattedTotalResults}{No description}

\item{searchTime}{No description}

\item{totalResults}{No description}
}
\value{
Search.searchInformation object
}
\description{
Search.searchInformation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Search functions: \code{\link{Search.queries}},
  \code{\link{Search.spelling}}, \code{\link{Search.url}},
  \code{\link{Search}}
}

