% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Search.spelling}
\alias{Search.spelling}
\title{Search.spelling Object}
\usage{
Search.spelling(correctedQuery = NULL, htmlCorrectedQuery = NULL)
}
\arguments{
\item{correctedQuery}{No description}

\item{htmlCorrectedQuery}{No description}
}
\value{
Search.spelling object
}
\description{
Search.spelling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Search functions: \code{\link{Search.queries}},
  \code{\link{Search.searchInformation}},
  \code{\link{Search.url}}, \code{\link{Search}}
}

