% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customsearch_objects.R
\name{Search.url}
\alias{Search.url}
\title{Search.url Object}
\usage{
Search.url(template = NULL, type = NULL)
}
\arguments{
\item{template}{No description}

\item{type}{No description}
}
\value{
Search.url object
}
\description{
Search.url Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Search functions: \code{\link{Search.queries}},
  \code{\link{Search.searchInformation}},
  \code{\link{Search.spelling}}, \code{\link{Search}}
}
